implementation of application:maned:read:notify consumer
