+++ doajtest/unit/event_consumers/test_application_assed_assigned_notify.py
@@ -2,31 +2,27 @@ from portality import models
 from portality import constants
 from portality.bll import exceptions
 from doajtest.helpers import DoajTestCase
-from portality.events.consumers.application_maned_ready_notify import ApplicationManedReadyNotify
+from portality.events.consumers.application_assed_assigned_notify import ApplicationAssedAssignedNotify
 from doajtest.fixtures import ApplicationFixtureFactory
 import time
 
 
-class TestApplicationManedReadyNotify(DoajTestCase):
+class TestApplicationAssedAssignedNotify(DoajTestCase):
     def setUp(self):
-        super(TestApplicationManedReadyNotify, self).setUp()
+        super(TestApplicationAssedAssignedNotify, self).setUp()
 
     def tearDown(self):
-        super(TestApplicationManedReadyNotify, self).tearDown()
+        super(TestApplicationAssedAssignedNotify, self).tearDown()
 
     def test_consumes(self):
-        event = models.Event(constants.EVENT_APPLICATION_STATUS, context={"application" : "2345", "old_status" : "in progress", "new_status": "ready"})
-        assert ApplicationManedReadyNotify.consumes(event)
-
-        event = models.Event(constants.EVENT_APPLICATION_STATUS,
-                             context={"application": "2345", "old_status": "ready", "new_status": "ready"})
-        assert not ApplicationManedReadyNotify.consumes(event)
+        event = models.Event(constants.EVENT_APPLICATION_ASSED_ASSIGNED, context={"application" : "2345"})
+        assert ApplicationAssedAssignedNotify.consumes(event)
 
         event = models.Event("test:event", context={"application" : "2345"})
-        assert not ApplicationManedReadyNotify.consumes(event)
+        assert not ApplicationAssedAssignedNotify.consumes(event)
 
-        event = models.Event(constants.EVENT_APPLICATION_STATUS)
-        assert not ApplicationManedReadyNotify.consumes(event)
+        event = models.Event(constants.EVENT_APPLICATION_ASSED_ASSIGNED)
+        assert not ApplicationAssedAssignedNotify.consumes(event)
 
     def test_consume_success(self):
         self._make_and_push_test_context("/")
@@ -41,7 +37,7 @@ class TestApplicationManedReadyNotify(DoajTestCase):
         acc.save(blocking=True)
 
         event = models.Event(constants.EVENT_APPLICATION_ASSED_ASSIGNED, context={"application" : app.id})
-        ApplicationManedReadyNotify.consume(event)
+        ApplicationAssedAssignedNotify.consume(event)
 
         time.sleep(2)
         ns = models.Notification.all()
@@ -49,7 +45,7 @@ class TestApplicationManedReadyNotify(DoajTestCase):
 
         n = ns[0]
         assert n.who == app.editor
-        assert n.created_by == ApplicationManedReadyNotify.ID
+        assert n.created_by == ApplicationAssedAssignedNotify.ID
         assert n.classification == constants.NOTIFICATION_CLASSIFICATION_ASSIGN
         assert n.message is not None
         assert n.action is not None
@@ -58,7 +54,7 @@ class TestApplicationManedReadyNotify(DoajTestCase):
     def test_consume_fail(self):
         event = models.Event(constants.EVENT_APPLICATION_ASSED_ASSIGNED, context={"application": "abcd"})
         with self.assertRaises(exceptions.NoSuchObjectException):
-            ApplicationManedReadyNotify.consume(event)
+            ApplicationAssedAssignedNotify.consume(event)
 
         source = ApplicationFixtureFactory.make_application_source()
         del source["admin"]["editor"]
@@ -67,4 +63,4 @@ class TestApplicationManedReadyNotify(DoajTestCase):
 
         event = models.Event(constants.EVENT_APPLICATION_ASSED_ASSIGNED, context={"application": app.id})
         with self.assertRaises(exceptions.NoSuchPropertyException):
-            ApplicationManedReadyNotify.consume(event)
+            ApplicationAssedAssignedNotify.consume(event)
