+++ doajtest/unit/event_consumers/test_application_maned_ready_notify.py
@@ -2,27 +2,31 @@ from portality import models
 from portality import constants
 from portality.bll import exceptions
 from doajtest.helpers import DoajTestCase
-from portality.events.consumers.application_assed_assigned_notify import ApplicationAssedAssignedNotify
+from portality.events.consumers.application_maned_ready_notify import ApplicationManedReadyNotify
 from doajtest.fixtures import ApplicationFixtureFactory
 import time
 
 
-class TestApplicationAssedAssignedNotify(DoajTestCase):
+class TestApplicationManedReadyNotify(DoajTestCase):
     def setUp(self):
-        super(TestApplicationAssedAssignedNotify, self).setUp()
+        super(TestApplicationManedReadyNotify, self).setUp()
 
     def tearDown(self):
-        super(TestApplicationAssedAssignedNotify, self).tearDown()
+        super(TestApplicationManedReadyNotify, self).tearDown()
 
     def test_consumes(self):
-        event = models.Event(constants.EVENT_APPLICATION_ASSED_ASSIGNED, context={"application" : "2345"})
-        assert ApplicationAssedAssignedNotify.consumes(event)
+        event = models.Event(constants.EVENT_APPLICATION_STATUS, context={"application" : "2345", "old_status" : "in progress", "new_status": "ready"})
+        assert ApplicationManedReadyNotify.consumes(event)
+
+        event = models.Event(constants.EVENT_APPLICATION_STATUS,
+                             context={"application": "2345", "old_status": "ready", "new_status": "ready"})
+        assert not ApplicationManedReadyNotify.consumes(event)
 
         event = models.Event("test:event", context={"application" : "2345"})
-        assert not ApplicationAssedAssignedNotify.consumes(event)
+        assert not ApplicationManedReadyNotify.consumes(event)
 
-        event = models.Event(constants.EVENT_APPLICATION_ASSED_ASSIGNED)
-        assert not ApplicationAssedAssignedNotify.consumes(event)
+        event = models.Event(constants.EVENT_APPLICATION_STATUS)
+        assert not ApplicationManedReadyNotify.consumes(event)
 
     def test_consume_success(self):
         self._make_and_push_test_context("/")
@@ -32,21 +36,26 @@ class TestApplicationAssedAssignedNotify(DoajTestCase):
         app.save()
 
         acc = models.Account()
-        acc.set_id(app.editor)
+        acc.set_id("maned")
         acc.set_email("test@example.com")
-        acc.save(blocking=True)
+        acc.save()
 
-        event = models.Event(constants.EVENT_APPLICATION_ASSED_ASSIGNED, context={"application" : app.id})
-        ApplicationAssedAssignedNotify.consume(event)
+        eg = models.EditorGroup()
+        eg.set_name(app.editor_group)
+        eg.set_maned(acc.id)
+        eg.save(blocking=True)
+
+        event = models.Event(constants.EVENT_APPLICATION_STATUS, context={"application" : app.id, "old_status": "in progress", "new_status": "ready"})
+        ApplicationManedReadyNotify.consume(event)
 
         time.sleep(2)
         ns = models.Notification.all()
         assert len(ns) == 1
 
         n = ns[0]
-        assert n.who == app.editor
-        assert n.created_by == ApplicationAssedAssignedNotify.ID
-        assert n.classification == constants.NOTIFICATION_CLASSIFICATION_ASSIGN
+        assert n.who == "maned"
+        assert n.created_by == ApplicationManedReadyNotify.ID
+        assert n.classification == constants.NOTIFICATION_CLASSIFICATION_STATUS_CHANGE
         assert n.message is not None
         assert n.action is not None
         assert not n.is_seen()
@@ -54,13 +63,5 @@ class TestApplicationAssedAssignedNotify(DoajTestCase):
     def test_consume_fail(self):
         event = models.Event(constants.EVENT_APPLICATION_ASSED_ASSIGNED, context={"application": "abcd"})
         with self.assertRaises(exceptions.NoSuchObjectException):
-            ApplicationAssedAssignedNotify.consume(event)
+            ApplicationManedReadyNotify.consume(event)
 
-        source = ApplicationFixtureFactory.make_application_source()
-        del source["admin"]["editor"]
-        app = models.Application(**source)
-        app.save(blocking=True)
-
-        event = models.Event(constants.EVENT_APPLICATION_ASSED_ASSIGNED, context={"application": app.id})
-        with self.assertRaises(exceptions.NoSuchPropertyException):
-            ApplicationAssedAssignedNotify.consume(event)
