+++ portality/templates/email/notification_email.jinja2
@@ -2,7 +2,9 @@
 Dear {{ user.name }},
 {% endif -%}
 
+
 {{ message }}
+
 
 {%- if action %}
 To take action based on this notification, follow this link: {{ url_root }}{{ action }}
